FE - Move Main Components Into the Pages Directory
