+++ client/src/components/common/devicesPageWrapper.tsx
@@ -2,7 +2,6 @@
 import React from 'react';
 
 // Custom
-import '../../App.scss';
 import Navbar from './Navbar';
 
 const DevicesPageWrapper = ({children}: { children: JSX.Element }) => {
